add data type 0
