+++ griblib/data0_test.go
@@ -1,0 +1,24 @@
+package griblib
+
+import (
+	"os"
+	"testing"
+)
+
+func Test_read0_integrationtest_file_hour0(t *testing.T) {
+	testFile, fileOpenErr := os.Open("integrationtestdata/template0.grb2")
+
+	if fileOpenErr != nil {
+		t.Fatal("Grib file for integration tests not found")
+	}
+	messages, messageParseErr := ReadMessages(testFile)
+
+	if messageParseErr != nil {
+		t.Fatal("Error reading messages: ", messageParseErr.Error())
+	}
+
+	if len(messages) != 2 {
+		t.Errorf("should have exactly 354 message in testfile, was %d", len(messages))
+	}
+
+}
